requirements.txt
fastapi
uvicorn
openai
google-api-python-client
google-auth-oauthlib
python-dotenv
requests
pydantic
python-dateutil
pymongo
dnspython
apscheduler
pytz
